paid-marketing-agents/requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
slack-sdk>=3.27.0
pyairtable>=2.3.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
